+++ src/index.ts
@@ -1,6 +1,6 @@
 import '../images/empty-img.png';
 import '../src/styles/index.css';
-import VendingMachineComponent from '../src/components/index';
+import VendingMachineComponent from './components/VendingMachineComponent';
 
 const hashRoute = window.location.hash.replace('#', '');
 const vendingMachineComponent = new VendingMachineComponent(hashRoute);
